requirements.txt
pytest
streamlit==1.31.1
langchain==0.1.8
langchain-community==0.0.21
langchain-core>=0.1.45
langchain-openai==0.0.6
mysql-connector-python==8.3.0
groq==0.4.1
langchain-groq==0.1.9
python-dotenv
pandas>=1.5.0
plotly>=5.0.0
numpy>=1.21.0